+++ pages/group/[id]/expense.tsx
@@ -1,0 +1,8 @@
+import { FunctionComponent } from "react";
+import { Container } from "../../../src/components/Container";
+
+const Expense: FunctionComponent = () => {
+  return <Container>hello add expense</Container>;
+};
+
+export default Expense;
